Debugging and features update. - As line height is now controlled through EPub-InkPlate options, CSS's   line-height attribute is no longer read. - Unzip class reorganized around the notion of active session to mitigate   potential mutex deadlocks. - Some new drawings icons added for eventual usage (waiting and BlueTooth icons). - Insure that page_locs task is disabled as the dir database is updated. - Adjusted PageLocsRetriever to be in synch with the main EPub viewer in regard   of the PageFormat parameters. - Partition usage has been modify to have the application uses the maximum   available space. That is in support of future features to be added. - As multi-columns augment the chance of having words that don't fit in shorter   lines. I have added a way to cut words in half, a kind of hyphenation.   This is not using any hyphenation database so the cut is done arbitrarily   in a way to get the words fit in the available space. When there is hyphens   already present in a word, they are used instead.

+++ src/controllers/ble_keypad.cpp
@@ -193,7 +193,7 @@
       if (param->scan_start_cmpl.status == ESP_BT_STATUS_SUCCESS) {
         LOG_W("=== BLE SCANNER RUNNING ===");
       } else {
-        LOG_E("Scan start failed: {}", param->scan_start_cmpl.status);
+        LOG_E("Scan start failed: {}", (int)param->scan_start_cmpl.status);
       }
       break;
 
@@ -232,7 +232,7 @@
 
     case ESP_GATTC_OPEN_EVT:
       if (param->open.status != ESP_GATT_OK) {
-        LOG_E("Failed to map GATT channel, error status: {}", param->open.status);
+        LOG_E("Failed to map GATT channel, error status: {}", (int)param->open.status);
         is_connecting = false;
       } else {
         gl_conn_id = param->open.conn_id;
